# swagger.json
{"openapi":"3.0.0","paths":{"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":""}},"tags":["Auth"]}},"/employees":{"post":{"operationId":"EmployeeController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateEmployeeDto"}}}},"responses":{"201":{"description":""}},"tags":["Employee"]},"get":{"operationId":"EmployeeController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Employee"]}},"/employees/{id}":{"get":{"operationId":"EmployeeController_findById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Employee"]},"patch":{"operationId":"EmployeeController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Employee"]},"delete":{"operationId":"EmployeeController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Employee"]}},"/employees/username/{username}":{"get":{"operationId":"EmployeeController_findByUsername","parameters":[{"name":"username","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Employee"]}},"/employees/code/{code}":{"get":{"operationId":"EmployeeController_findByCode","parameters":[{"name":"code","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Employee"]}},"/clients":{"post":{"operationId":"ClientController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateClientDto"}}}},"responses":{"201":{"description":""}},"tags":["Client"]},"get":{"operationId":"ClientController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Client"]}},"/clients/{id}":{"get":{"operationId":"ClientController_findById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Client"]},"patch":{"operationId":"ClientController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Client"]},"delete":{"operationId":"ClientController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Client"]}},"/clients/{id}/installments":{"get":{"operationId":"ClientController_getClientInstallments","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Client"]}},"/sales":{"post":{"operationId":"SaleController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSaleDto"}}}},"responses":{"201":{"description":""}},"tags":["Sale"]},"get":{"operationId":"SaleController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Sale"]}},"/sales/{id}":{"get":{"operationId":"SaleController_findById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Sale"]},"patch":{"operationId":"SaleController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Sale"]},"delete":{"operationId":"SaleController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Sale"]}},"/sales/{id}/installment":{"patch":{"operationId":"SaleController_updateInstallmentStatus","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InstallmentDto"}}}},"responses":{"200":{"description":""}},"tags":["Sale"]}},"/products":{"post":{"operationId":"ProductController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProductDto"}}}},"responses":{"201":{"description":""}},"tags":["Product"]},"get":{"operationId":"ProductController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Product"]}},"/products/{id}":{"get":{"operationId":"ProductController_findById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Product"]},"patch":{"operationId":"ProductController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Product"]},"delete":{"operationId":"ProductController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Product"]}},"/invoices":{"post":{"operationId":"InvoiceController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateInvoiceDto"}}}},"responses":{"201":{"description":""}},"tags":["Invoice"]},"get":{"operationId":"InvoiceController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Invoice"]}},"/invoices/{id}":{"get":{"operationId":"InvoiceController_findById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Invoice"]},"patch":{"operationId":"InvoiceController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Invoice"]},"delete":{"operationId":"InvoiceController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Invoice"]}},"/suppliers":{"post":{"operationId":"SupplierController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSupplierDto"}}}},"responses":{"201":{"description":""}},"tags":["Supplier"]},"get":{"operationId":"SupplierController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Supplier"]}},"/suppliers/{id}":{"get":{"operationId":"SupplierController_findById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Supplier"]},"patch":{"operationId":"SupplierController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSupplierDto"}}}},"responses":{"200":{"description":""}},"tags":["Supplier"]},"delete":{"operationId":"SupplierController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Supplier"]}},"/categories":{"post":{"operationId":"CategoryController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCategoryDto"}}}},"responses":{"201":{"description":""}},"tags":["Category"]},"get":{"operationId":"CategoryController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Category"]}},"/categories/{id}":{"get":{"operationId":"CategoryController_findById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Category"]},"patch":{"operationId":"CategoryController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Category"]},"delete":{"operationId":"CategoryController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Category"]}}},"info":{"title":"Documentação API MSR","description":"Bem-vindo à documentação da API do sistema de gestão MSR. Aqui você encontrará informações detalhadas sobre os endpoints disponíveis, incluindo métodos, parâmetros, e exemplos de uso. Esta API é projetada para gerenciar recursos do sistema MSR de forma eficiente e segura.","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"LoginDto":{"type":"object","properties":{"username":{"type":"string","example":"username123","description":"Username escolhido no cadastro do funcionário"},"password":{"type":"string","example":"123456","description":"Senha escolhida no cadastro do funcionário"}},"required":["username","password"]},"AddressDto":{"type":"object","properties":{"street":{"type":"string","example":"Rua das Flores","description":"Nome da rua do endereço"},"number":{"type":"string","example":"123","description":"Número do endereço"},"complement":{"type":"string","example":"Apto 101","description":"Complemento do endereço (opcional)"},"city":{"type":"string","example":"São Paulo","description":"Cidade do endereço"},"state":{"type":"string","example":"SP","description":"Estado do endereço"},"postalCode":{"type":"string","example":"12345-678","description":"CEP do endereço"},"country":{"type":"string","example":"Brasil","description":"País do endereço (opcional)"}},"required":["street","number","city","state","postalCode"]},"CreateEmployeeDto":{"type":"object","properties":{"name":{"type":"string","example":"John Doe","description":"Nome completo do funcionário"},"username":{"type":"string","example":"johndoe","description":"Nome de usuário único para login"},"password":{"type":"string","example":"password123","description":"Senha do funcionário para login"},"cpf":{"type":"string","example":"123.456.789-00","description":"CPF do funcionário"},"phone":{"type":"string","example":"11987654321","description":"Número de telefone do funcionário (máximo 11 caracteres)"},"email":{"type":"string","example":"johndoe@example.com","description":"Endereço de e-mail do funcionário"},"isAdmin":{"type":"boolean","example":false,"description":"Indica se o funcionário tem privilégios de administrador"},"address":{"description":"Endereço do funcionário","allOf":[{"$ref":"#/components/schemas/AddressDto"}]}},"required":["name","username","password","cpf","phone"]},"CreateClientDto":{"type":"object","properties":{"name":{"type":"string","example":"John Doe","description":"Nome completo do cliente"},"birthDate":{"format":"date-time","type":"string","example":"1990-01-01","description":"Data de nascimento do cliente"},"cpf":{"type":"string","example":"123.456.789-00","description":"CPF do cliente"},"rg":{"type":"string","example":"12.345.678-9","description":"RG do cliente"},"phone":{"type":"string","example":"11987654321","description":"Número de telefone do cliente (máximo 11 caracteres)"},"email":{"type":"string","example":"johndoe@example.com","description":"Endereço de e-mail do cliente (opcional)"},"purchaseLimit":{"type":"number","example":10000,"description":"Limite de compras do cliente (opcional)"},"notes":{"type":"string","example":"Cliente VIP","description":"Notas adicionais sobre o cliente (opcional)"},"createdByEmployee":{"type":"string","example":"644a1f4e5f1b2c3d4e5f6a7b","description":"ID do funcionário que cadastrou o cliente (opcional)"},"fathersName":{"type":"string","example":"John Senior","description":"Nome do pai do cliente (opcional)"},"mothersName":{"type":"string","example":"Jane Doe","description":"Nome da mãe do cliente (opcional)"},"peopleAuthorized":{"type":"string","example":"Maria Silva, Pedro Santos","description":"Pessoas autorizadas a realizar compras em nome do cliente (opcional)"},"timesCharged":{"type":"number","example":3,"description":"Número de vezes que o cliente foi cobrado (opcional)"},"spcInclusionDate":{"format":"date-time","type":"string","example":"2025-01-01","description":"Data de inclusão do cliente no SPC (opcional)"},"spcExclusionDate":{"format":"date-time","type":"string","example":"2025-06-01","description":"Data de exclusão do cliente do SPC (opcional)"},"spcExclusionReason":{"type":"string","example":"Dívida quitada","description":"Motivo da exclusão do cliente do SPC (opcional)"},"address":{"description":"Endereço do cliente (opcional)","allOf":[{"$ref":"#/components/schemas/AddressDto"}]}},"required":["name","birthDate","cpf","rg","phone"]},"SaleProductDto":{"type":"object","properties":{"_id":{"type":"string","example":"644a1f4e5f1b2c3d4e5f6a7b","description":"ID do produto"},"code":{"type":"number","example":12345,"description":"Código único do produto"},"name":{"type":"string","example":"Produto Exemplo","description":"Nome do produto"},"unitPrice":{"type":"number","example":50.25,"description":"Preço unitário do produto"},"totalPrice":{"type":"number","example":100.5,"description":"Preço total do produto (quantidade x preço unitário)"},"quantity":{"type":"number","example":2,"description":"Quantidade do produto na venda"}},"required":["_id","code","name","unitPrice","totalPrice","quantity"]},"Installment":{"type":"object","properties":{}},"CreateSaleDto":{"type":"object","properties":{"products":{"description":"Lista de produtos incluídos na venda","type":"array","items":{"$ref":"#/components/schemas/SaleProductDto"}},"clientId":{"type":"string","example":"644a1f4e5f1b2c3d4e5f6a7b","description":"ID do cliente associado à venda (opcional)"},"openedByEmployee":{"type":"string","example":"644a1f4e5f1b2c3d4e5f6a7b","description":"ID do funcionário que abriu a venda"},"total":{"type":"number","example":150.75,"description":"Valor total da venda"},"notes":{"type":"string","example":"Venda realizada com desconto especial","description":"Notas adicionais sobre a venda (opcional)"},"status":{"type":"string","example":"open","description":"Status da venda (aberta ou fechada)","enum":["open","closed"]},"paymentMethod":{"type":"string","example":"credit","description":"Método de pagamento utilizado na venda","enum":["credit","debit","cash","pix","bankTransfer","client-account"]},"installments":{"example":[],"description":"Parcelas associadas à venda (opcional)","type":"array","items":{"$ref":"#/components/schemas/Installment"}}},"required":["products","openedByEmployee","total"]},"InstallmentDto":{"type":"object","properties":{"_id":{"type":"string","example":"644a1f4e5f1b2c3d4e5f6a7b","description":"ID da parcela da venda"},"dueDate":{"format":"date-time","type":"string","example":"2025-05-01","description":"Data de vencimento da parcela"},"amount":{"type":"number","example":150.75,"description":"Valor da parcela"},"status":{"type":"string","example":"pending","description":"Status da parcela","enum":["pending","paid","overdue"]},"paymentDate":{"format":"date-time","type":"string","example":"2025-05-01","description":"Data que a parcela foi paga"},"paymentMethod":{"type":"string","example":"credit","description":"Método de pagamento utilizado na venda","enum":["credit","debit","cash","pix","bankTransfer"]}},"required":["_id","dueDate","amount"]},"CreateProductDto":{"type":"object","properties":{"name":{"type":"string","example":"Caneta Azul","description":"Nome do produto"},"description":{"type":"string","example":"Caneta esferográfica azul com tinta de alta qualidade","description":"Descrição detalhada do produto (opcional)"},"price":{"type":"number","example":2.5,"description":"Preço unitário do produto"},"stock":{"type":"number","example":100,"description":"Quantidade em estoque do produto"},"supplierId":{"type":"string","example":"644a1f4e5f1b2c3d4e5f6a7b","description":"ID do fornecedor associado ao produto (opcional)"},"categories":{"example":["644a1f4e5f1b2c3d4e5f6a7b","644a1f4e5f1b2c3d4e5f6a7c"],"description":"Lista de IDs das categorias associadas ao produto (opcional)","type":"array","items":{"type":"string"}}},"required":["name","price","stock"]},"CreateInvoiceDto":{"type":"object","properties":{"supplierId":{"type":"string","example":"644a1f4e5f1b2c3d4e5f6a7b","description":"ID do fornecedor associado à fatura (opcional)"},"issueDate":{"format":"date-time","type":"string","example":"2025-04-01","description":"Data de emissão da fatura (opcional)"},"dueDate":{"format":"date-time","type":"string","example":"2025-05-01","description":"Data de vencimento da fatura"},"totalAmount":{"type":"number","example":1500.75,"description":"Valor total da fatura"},"installments":{"description":"Lista de parcelas associadas à fatura (opcional)","type":"array","items":{"$ref":"#/components/schemas/InstallmentDto"}},"status":{"type":"string","example":"pending","description":"Status da fatura (opcional)","enum":["pending","paid","overdue"]},"notes":{"type":"string","example":"Fatura referente à compra de materiais de escritório","description":"Notas adicionais sobre a fatura (opcional)"},"description":{"type":"string","example":"Compra de materiais de escritório","description":"Descrição da fatura (opcional)"}},"required":["dueDate","totalAmount"]},"CreateSupplierDto":{"type":"object","properties":{"name":{"type":"string","example":"Fornecedor ABC","description":"Nome do fornecedor"},"cnpj":{"type":"string","example":"12.345.678/0001-99","description":"CNPJ do fornecedor"},"phone":{"type":"string","example":"11987654321","description":"Número de telefone do fornecedor (opcional)"},"email":{"type":"string","example":"contato@fornecedorabc.com","description":"Endereço de e-mail do fornecedor (opcional)"},"contactPerson":{"type":"string","example":"João Silva","description":"Nome da pessoa de contato no fornecedor (opcional)"},"notes":{"type":"string","example":"Fornecedor confiável com entrega rápida","description":"Notas adicionais sobre o fornecedor (opcional)"},"address":{"description":"Endereço do fornecedor (opcional)","allOf":[{"$ref":"#/components/schemas/AddressDto"}]}},"required":["name","cnpj"]},"UpdateSupplierDto":{"type":"object","properties":{}},"CreateCategoryDto":{"type":"object","properties":{"name":{"type":"string","example":"Eletrônicos","description":"Nome da categoria"},"description":{"type":"string","example":"Categoria para produtos eletrônicos, como celulares e laptops","description":"Descrição detalhada da categoria (opcional)"}},"required":["name"]}}}}